uml: tests/md: Skip test-inspect-fstab.sh since it needs qcow2.

+++ tests/md/test-inspect-fstab.sh
@@ -25,6 +25,11 @@ export LANG=C
 guestfish=../../fish/guestfish
 canonical="sed -r s,/dev/[abce-ln-z]+d,/dev/sd,g"
 
+if [ "$($guestfish get-backend)" = "uml" ]; then
+    echo "$0: skipping test because uml backend does not support qcow2"
+    exit 77
+fi
+
 rm -f inspect-fstab-1.qcow2 inspect-fstab.fstab inspect-fstab.output
 
 # Start with the regular (good) fedora image, modify /etc/fstab
